requirements.txt
Flask
Flask-SocketIO
Flask-Login
Flask-Mail
SQLAlchemy
eventlet
python-dotenv
